Refactoring cartelle
