+++ app/src/main/java/it/unimib/brewday/ui/Callback.java
@@ -1,6 +1,7 @@
-package it.unimib.brewday;
+package it.unimib.brewday.ui;
 
 
+import it.unimib.brewday.model.Risultato;
 
 public interface Callback {
     void onComplete(Risultato risultato);
