+++ app/src/main/java/it/unimib/brewday/ui/MainActivity.java
@@ -1,8 +1,10 @@
-package it.unimib.brewday;
+package it.unimib.brewday.ui;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import it.unimib.brewday.R;
 
 public class MainActivity extends AppCompatActivity {
