+++ app/src/main/res/layout/activity_main.xml
@@ -4,11 +4,11 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".ui.MainActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/ActivityTest"
-        android:name="it.unimib.brewday.IngredientiFragment"
+        android:name="it.unimib.brewday.ui.IngredientiFragment"
         android:layout_height="match_parent"
         android:layout_width="match_parent"
         tools:layout="@layout/fragment_ingredienti" />
